requirements.txt
streamlit >= 1.10.0
panda >= 0.3.1
numpy >= 1.19.4